Migrate to net/http middleware
Instead of just being for chi, we can take this opportunity to make it
clear this works for any net/http compatible router.

We can also make sure we've got test coverage for both Chi and Gorilla,
the routers that right now would be using this middleware, which are
tracked in subpackages to avoid polluting the dependency tree.

As this is now a multi-module project, we need to make sure our `make`
tasks handle this.