fix: return HTTP 405 when request method mismatch (#25)
When a route is correctly matched, but the HTTP method used does not, we
should return an HTTP 405 Method Not Allowed.

However, in our previous implementation, we would always return an HTTP
404 Not Found, which is incorrect.

Instead, we can match on whether we have received an
`ErrMethodNotAllowed` and if so, return the correct HTTP 405.